# specs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Stellar Scout API",
		"version": "1.6.1",
		"summary": "Agent-native data layer for the Stellar ecosystem — projects, builders, hackathons, SCF, audits, research.",
...
		"contact": {
			"name": "Stellar Light",
			"url": "https://stellarlight.xyz/scout",
			"email": "support@stellarlight.xyz"
		},
		"license": {
			"name": "MIT",
			"identifier": "MIT"
		}
	},
	"servers": [
		{
			"url": "https://stellarlight.xyz",
			"description": "Production"
		}
	],
...
	"paths": {
...
		"/api/partners/onboard": {
			"post": {
				"operationId": "partnerOnboard",
				"tags": [
					"Partners"
				],
				"summary": "AI onboarding helpers: interview chat + profile extraction",
				"description": "Two modes for the get-listed flow. `{mode:'chat', messages}` → interview-style reply for a company describing itself. `{mode:'extract', messages}` → structures the transcript into partner-profile `fields` (tagline, services, sectors, regions, contact…; null where nothing was stated — nothing invented). 503 `unavailable:true` without AI configured. **Use when:** turning a get-listed conversation into structured profile fields (then submit via POST /api/partners/submit-listing). **Not for:** finding partners → /api/partners/match or /assistant.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"mode",
									"messages"
								],
								"properties": {
									"mode": {
										"type": "string",
										"enum": [
											"chat",
											"extract"
										]
									},
									"messages": {
										"type": "array",
										"items": {
											"type": "object",
											"required": [
												"role",
												"content"
											],
											"properties": {
												"role": {
													"type": "string",
													"enum": [
														"user",
														"assistant"
													]
												},
												"content": {
													"type": "string"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "`{reply}` (chat mode) or `{fields}` (extract mode)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					},
					"429": {
						"description": "Rate limited"
					},
					"503": {
						"description": "AI backend unavailable",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}